Revert "feat: Refactor pages to use single Page component (#313)"
This reverts commit f636a06ea6c4734f8551a85ae1642a7636c53b57.
