+++ src/pages/calendar/index.tsx
@@ -3,7 +3,7 @@ import { useSearchParams } from 'react-router-dom';
 
 import { Term } from 'lib/fetchers';
 
-import { Page } from 'common/page/Page';
+import { SidebarTemplate } from 'common/sidebar/';
 
 import { Landing } from 'pages/home/containers/Landing';
 
@@ -16,8 +16,8 @@ export function Calendar(): JSX.Element {
   const pid = searchParams.get('pid');
 
   return (
-    <Page title="Calendar" hasSearchableSidebar>
+    <SidebarTemplate title="Calendar" term={term as Term}>
       {pid ? <Content term={term as Term} /> : <Landing />}
-    </Page>
+    </SidebarTemplate>
   );
 }
