+++ src/pages/registration/containers/RegistrationContainer.tsx
@@ -1,12 +1,13 @@
-import { Box, Container, Divider, Heading } from '@chakra-ui/layout';
-import { Center, Spinner, Text } from '@chakra-ui/react';
+import { Box, Container, Divider, Flex, Heading } from '@chakra-ui/layout';
+import { Text } from '@chakra-ui/react';
+import { Helmet } from 'react-helmet';
 import { useParams } from 'react-router';
 
 import { useDarkMode } from 'lib/hooks/useDarkMode';
 import { useSavedCourses } from 'lib/hooks/useSavedCourses';
 import { getReadableTerm } from 'lib/utils/terms';
 
-import { Page } from 'common/page/Page';
+import { Header } from 'common/header';
 
 import { RegistrationHeading } from '../components/RegistrationHeading';
 
@@ -17,41 +18,36 @@ export function RegistrationContainer(): JSX.Element | null {
   const { courses } = useSavedCourses();
   const mode = useDarkMode();
 
-  // to avoid erroring out if term is not provided in URL
-  // term is eventually filled in but need to avoid initial error
-  if (!term)
-    return (
-      <Page title="Loading registration data...">
-        <Center height="100%" mt="10">
-          <Spinner size="xl" />
-        </Center>
-      </Page>
-    );
-
   return (
-    <Page title={`${getReadableTerm(term)} · Registration`}>
-      <Box maxW={{ base: '35rem', md: '65rem' }} textAlign="center">
-        <RegistrationHeading />
-        {courses.filter((course) => course.term === term).length > 0 ? (
-          courses
-            .filter((course) => course.term === term)
-            .map((course) => {
-              return <CourseContainer course={course} />;
-            })
-        ) : (
-          <>
-            <Divider my="4" />
-            <Container alignItems="center" maxW="container.xl">
-              <Heading size="md" color={mode('gray', 'dark.header')}>
-                Unable to find saved courses for{' '}
-                <Text as="span" color={mode('black', 'white')}>
-                  {getReadableTerm(term)}
-                </Text>
-              </Heading>
-            </Container>
-          </>
-        )}
-      </Box>
-    </Page>
+    <Flex h="100vh" direction="column" overflowX="hidden" overflowY="hidden">
+      <Helmet>
+        <title>{`${getReadableTerm(term)} · Registration`}</title>
+      </Helmet>
+      <Header />
+      <Flex width="100%" pt="1.25rem" direction="column" alignItems="center" overflowY="auto">
+        <Box maxW={{ base: '35rem', md: '65rem' }} textAlign="center">
+          <RegistrationHeading />
+          {courses.filter((course) => course.term === term).length > 0 ? (
+            courses
+              .filter((course) => course.term === term)
+              .map((course) => {
+                return <CourseContainer course={course} />;
+              })
+          ) : (
+            <>
+              <Divider my="4" />
+              <Container alignItems="center" maxW="container.xl">
+                <Heading size="md" color={mode('gray', 'dark.header')}>
+                  Unable to find saved courses for{' '}
+                  <Text as="span" color={mode('black', 'white')}>
+                    {getReadableTerm(term)}
+                  </Text>
+                </Heading>
+              </Container>
+            </>
+          )}
+        </Box>
+      </Flex>
+    </Flex>
   );
 }
