+++ src/pages/scheduler/index.tsx
@@ -1,7 +1,9 @@
 import { Flex } from '@chakra-ui/layout';
 import { useParams } from 'react-router';
 
-import { Page } from 'common/page/Page';
+import { Term } from 'lib/fetchers';
+
+import { SidebarTemplate } from 'common/sidebar/';
 
 import { SchedulerSidebar } from './components/SchedulerSidebar';
 import { SchedulerContainer } from './containers/SchedulerContainer';
@@ -10,11 +12,11 @@ export function Scheduler(): JSX.Element {
   const { term } = useParams();
 
   return (
-    <Page title="Scheduler" hasSearchableSidebar>
+    <SidebarTemplate title="Scheduler" term={term as Term}>
       <Flex flexGrow={1}>
         <SchedulerContainer />
         <SchedulerSidebar term={term} />
       </Flex>
-    </Page>
+    </SidebarTemplate>
   );
 }
